+++ src/yt_game_genesis.c
@@ -28,7 +28,7 @@ yt_genesis_confirmation_prompt(const uint8_t *trader, size_t trader_length,
 }
 
 bool
-yt_genesis_insufficient_rows(float required, float owned,
+yt_genesis_insufficient_rows(float required, uint16_t owned,
     uint8_t *first, size_t first_capacity, size_t *first_length,
     uint8_t *second, size_t second_capacity, size_t *second_length)
 {
@@ -39,7 +39,7 @@ yt_genesis_insufficient_rows(float required, float owned,
 	    "enough to initiate Genesis. You are";
 	static const uint8_t second_suffix[] =
 	    " short of fulfilling the prophesy.";
-	volatile float shortfall = required - owned;
+	volatile float shortfall = required - (float)owned;
 	char required_text[64];
 	char shortfall_text[64];
 	int required_length;
