+++ src/yt_maint.h
@@ -133,7 +133,7 @@ bool yt_maintenance_xannor_should_retarget(float group_location,
 bool yt_maintenance_xannor_route_complete(float group_location,
     int target_sector);
 bool yt_maintenance_xannor_bypass_initial_arrival(int group_number,
-    float group_location, float top_player_target);
+	float group_location, uint16_t top_player_target);
 bool yt_maintenance_xannor_should_attack_hunt_player(int group_number,
     int hunt_player);
 bool yt_maintenance_xannor_advance_group(int group_number,
